Makefile: cleanup golang from protobuf
The GOPATH variables were not quite right, as the vendor source did not
have ./src/

This adds the compiled golang source to the repo, so folks doing `go
get ...` won't have to bootstrap the tools needed.
The change made to the protobuf go_package was due to it producing that full
path in the `--go_out=./` directory.

Also, making the targets/recipes where the dependencies only run if
needed.

Signed-off-by: Vincent Batts <vbatts@hashbangbash.com>