Filter expression parser implementation.
The following change introduces an Antlr4 grammar that defines the filter syntax.
The grammar supports logical operations and sequences of restriction expressions,
with the ability to reference object fields and invoke functions supported by the
API.

The syntax is identical to that supported by Google Cloud Logging Advanced Filters
with one notable exception: it can accurately parse statements of the form:

    a[b]

as indexing operations on a variable. This is a known limitation of Advanced
Filters which is not present here.

Parsed filters are output in the Common Expression Language abstract syntax
representation. A key difference between CEL and list filters is that list filters
may treated barewords in the expression as unquoted text. Within CEL, these
barewords would be flagged as identifiers. Whether the value is a string or
identifier is a contextual decision best left to the filter evaluation step.