Use subtests in occurrence_test
Updating each test case to run as a subtest makes the output clearer.
This is most apparent in test cases that log errors; when all cases
run as a single test, the output contains the log statements from
all cases.

I came across this while making some changes that required updating
these tests; submitting as a separate PR as it's completely separate.