Add tests on permission checking
Add some tests to validate the permission checks on UpdateOccurrence
operations. To support the tests, add a allowListAuth test implementation
of Auth that just stores a list of allowed project/permission pairs.

Two of the tests are documenting incorrect behavior that will be fixed in
a later commit. Namely:
(1) Updating the NoteName should require NotesAttachOccurrence on the old
NoteName (as this is effectively a detach operation).
(2) When excluding the NoteName from the update using an UpdateMask, the
permission check should be performed against the existing NoteName, not
whatever was passed in (and which will be ignored).